requirements.txt
numpy
scipy
tqdm
pillow
urllib3
kagglehub
matplotlib
scikit-image
dlutils==0.0.2
torch >= 0.4.0
torchvision
torchsummary
pytorch-ignite